Refactored Phase 7: Extracted file operations into tools/files.py
- Moved all file operation tools to src/slack_mcp/tools/files.py
- Added aiohttp import for file download functionality
- Updated server to import and register file tools module
- Fixed upload_file to use files_upload instead of files_upload_v2
- Fixed parameter handling for file uploads (using channel IDs)
- Removed duplicate file tool definitions from main server file
- All file operations tested and working correctly
